chore: spotbug fixes

+++ runwaysdk-server/src/main/java/com/runwaysdk/vault/VaultCommand.java
@@ -3,18 +3,18 @@
  *
  * This file is part of Runway SDK(tm).
  *
- * Runway SDK(tm) is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as
- * published by the Free Software Foundation, either version 3 of the
- * License, or (at your option) any later version.
+ * Runway SDK(tm) is free software: you can redistribute it and/or modify it
+ * under the terms of the GNU Lesser General Public License as published by the
+ * Free Software Foundation, either version 3 of the License, or (at your
+ * option) any later version.
  *
- * Runway SDK(tm) is distributed in the hope that it will be useful, but
- * WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Lesser General Public License for more details.
+ * Runway SDK(tm) is distributed in the hope that it will be useful, but WITHOUT
+ * ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS
+ * FOR A PARTICULAR PURPOSE. See the GNU Lesser General Public License for more
+ * details.
  *
- * You should have received a copy of the GNU Lesser General Public
- * License along with Runway SDK(tm).  If not, see <http://www.gnu.org/licenses/>.
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with Runway SDK(tm). If not, see <http://www.gnu.org/licenses/>.
  */
 package com.runwaysdk.vault;
 
@@ -24,7 +24,6 @@ import java.io.IOException;
 import com.runwaysdk.SystemException;
 import com.runwaysdk.dataaccess.Command;
 import com.runwaysdk.util.FileIO;
-
 
 public class VaultCommand implements Command
 {
@@ -44,17 +43,22 @@ public class VaultCommand implements Command
 
   /**
    * Recursivley deletes all the files and sub directories of a given directory
-   * @param directory The directory
+   * 
+   * @param directory
+   *          The directory
    */
   private static void deleteDirectory(File directory)
   {
-    if(directory.isDirectory())
+    if (directory.isDirectory())
     {
       File[] files = directory.listFiles();
 
-      for(File file : files)
+      if (files != null)
       {
-        deleteDirectory(file);
+        for (File file : files)
+        {
+          deleteDirectory(file);
+        }
       }
     }
 
@@ -92,5 +96,7 @@ public class VaultCommand implements Command
    * transaction, this action is performed at the end of the transaction.
    */
   @Override
-  public void doFinally(){}
+  public void doFinally()
+  {
+  }
 }
